+++ examples/typescript/fullstack/anyspend/apps/client/index.html
@@ -2,9 +2,11 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/x-icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>AnySpend Fullstack Demo</title>
+    <meta name="description" content="Pay with any token using the x402 payment protocol - Demo application for EVM chains and Solana" />
+    <meta name="theme-color" content="#5B21B6" />
+    <title>AnySpend x402 - Pay with Any Token</title>
   </head>
   <body>
     <div id="root"></div>
