+++ examples/typescript/fullstack/anyspend/apps/client/package.json
@@ -5,23 +5,29 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
+    "start": "vite preview --host --port ${PORT:-4173}",
     "preview": "vite preview",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "format": "prettier --write \"src/**/*.{ts,tsx,css}\""
   },
   "dependencies": {
+    "@b3dotfun/anyspend-x402": "^0.2.4",
+    "@b3dotfun/anyspend-x402-fetch": "^0.2.4",
+    "@b3dotfun/anyspend-x402-solana-wallet-adapter": "workspace:^",
+    "@b3dotfun/anyspend-x402-token-compat": "^0.1.2",
+    "@solana/wallet-adapter-react": "^0.15.39",
+    "@solana/wallet-adapter-react-ui": "^0.9.39",
+    "@solana/wallet-adapter-wallets": "^0.19.37",
+    "@solana/web3.js": "^1.98.4",
+    "@tanstack/react-query": "^5.62.15",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "viem": "^2.21.26",
-    "wagmi": "^2.15.6",
-    "@tanstack/react-query": "^5.62.15",
-    "@b3dotfun/anyspend-x402": "workspace:*",
-    "@b3dotfun/anyspend-x402-fetch": "workspace:*",
-    "@b3dotfun/anyspend-x402-token-compat": "workspace:*"
+    "wagmi": "^2.15.6"
   },
   "devDependencies": {
-    "@types/react": "^18.2.43",
-    "@types/react-dom": "^18.2.17",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
     "@typescript-eslint/parser": "^6.14.0",
     "@vitejs/plugin-react": "^4.2.1",
@@ -30,6 +36,6 @@
     "eslint-plugin-react-refresh": "^0.4.5",
     "prettier": "^3.1.1",
     "typescript": "^5.2.2",
-    "vite": "^5.0.8"
+    "vite": "^7.1.12"
   }
 }
