+++ examples/typescript/fullstack/anyspend/apps/client/src/main.tsx
@@ -1,18 +1,21 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { WagmiProvider } from "wagmi";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import App from "./App.tsx";
 import { config } from "./wagmi.config";
+import { SolanaWalletProvider } from "./SolanaWalletProvider";
+import { WagmiProviderWrapper } from "./WagmiProviderWrapper";
 
 const queryClient = new QueryClient();
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <WagmiProvider config={config}>
+    <WagmiProviderWrapper config={config}>
       <QueryClientProvider client={queryClient}>
-        <App />
+        <SolanaWalletProvider>
+          <App />
+        </SolanaWalletProvider>
       </QueryClientProvider>
-    </WagmiProvider>
+    </WagmiProviderWrapper>
   </React.StrictMode>,
 );
