+++ examples/typescript/fullstack/anyspend/apps/client/vite.config.ts
@@ -7,4 +7,10 @@ export default defineConfig({
   server: {
     port: 3000,
   },
+  preview: {
+    allowedHosts: [
+      'anyspend-x402-client-production.up.railway.app',
+      'x402-demo.anyspend.com'
+    ]
+  }
 });
