+++ examples/typescript/fullstack/anyspend/apps/server/index.ts
@@ -1,7 +1,8 @@
-import { paymentMiddleware } from "@b3dotfun/anyspend-x402-express";
+import { paymentMiddleware, SolanaAddress } from "@b3dotfun/anyspend-x402-express";
 import cors from "cors";
 import dotenv from "dotenv";
 import express, { Request, Response } from "express";
+import http from "http";
 import { Address } from "viem";
 
 // Load environment variables
@@ -29,7 +30,7 @@ app.use(express.json());
 const PAYTO_ADDRESS =
   (process.env.PAYTO_ADDRESS as Address) || "0xB3B32F9f8827D4634fE7d973Fa1034Ec9fdDB3B3";
 const NETWORK = (process.env.NETWORK as "base-sepolia" | "base") || "base";
-const PAYMENT_AMOUNT_USD = process.env.PAYMENT_AMOUNT_USD || "100000000"; // Default 100 USDC (100 * 10^6)
+const PAYMENT_AMOUNT_USD = process.env.PAYMENT_AMOUNT_USD || "1000"; // Default 0.001 USDC (0.001 * 10^6)
 const FACILITATOR_URL = (process.env.FACILITATOR_URL ||
   "https://mainnet.anyspend.com/x402") as `${string}://${string}`;
 
@@ -44,12 +45,13 @@ const COINGECKO_API_KEY = process.env.COINGECKO_API_KEY!;
 const SIMDUNE_API_KEY = process.env.SIMDUNE_API_KEY!;
 
 // Solana Configuration
-const SOLANA_RPC_URL = process.env.SOLANA_RPC_URL || "https://api.mainnet-beta.solana.com";
+const SOLANA_RPC_URL = process.env.SOLANA_RPC_URL || "https://solana-rpc.publicnode.com";
 const SOLANA_USDC_MINT =
   process.env.SOLANA_USDC_MINT || "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v";
 const SOLANA_PAYMENT_AMOUNT = process.env.SOLANA_PAYMENT_AMOUNT || "10000"; // 0.01 USDC (6 decimals)
 const SOLANA_PAYTO_ADDRESS =
-  process.env.SOLANA_PAYTO_ADDRESS || "8Bw4C9cgFvMSsH3bdqJ1hpWPNdxknrawBPGzrHuYZR32";
+  (process.env.SOLANA_PAYTO_ADDRESS as SolanaAddress) ||
+  ("8Bw4C9cgFvMSsH3bdqJ1hpWPNdxknrawBPGzrHuYZR32" as SolanaAddress);
 
 // Apply payment middleware to protected routes
 app.use(
@@ -243,21 +245,18 @@ app.post("/api/btc", async (req: Request, res: Response) => {
  */
 app.post("/api/solana/premium", async (req: Request, res: Response) => {
   try {
+    // Fetch SOL price history from CoinGecko
+    const solData = await fetchSolPriceHistory();
+
     return res.json({
       success: true,
-      data: {
-        message: "Payment received successfully on Solana!",
-        network: "solana",
-        paymentToken: "USDC",
-        amount: SOLANA_PAYMENT_AMOUNT,
-        timestamp: new Date().toISOString(),
-      },
+      data: solData,
     });
   } catch (error) {
-    console.error("Error processing Solana payment:", error);
+    console.error("Error fetching SOL data:", error);
     return res.status(500).json({
       success: false,
-      error: error instanceof Error ? error.message : "Failed to process Solana payment",
+      error: error instanceof Error ? error.message : "Failed to fetch SOL data",
     });
   }
 });
@@ -545,6 +544,81 @@ async function fetchBtcPriceHistory() {
   }
 }
 
+/**
+ * Fetch SOL price history from CoinGecko
+ *
+ * @returns Promise with SOL price history data
+ */
+async function fetchSolPriceHistory() {
+  try {
+    // Fetch SOL OHLCV data for the last 24 hours (minute intervals)
+    const url =
+      "https://pro-api.coingecko.com/api/v3/coins/solana/ohlc?vs_currency=usd&days=1&precision=2";
+
+    console.log("Fetching SOL price history from CoinGecko...");
+
+    const response = await fetch(url, {
+      headers: {
+        "x-cg-pro-api-key": COINGECKO_API_KEY,
+      },
+    });
+
+    if (!response.ok) {
+      const errorText = await response.text();
+      console.error("CoinGecko API error:", response.status, errorText);
+      throw new Error(`CoinGecko API error: ${response.status}`);
+    }
+
+    const data = (await response.json()) as number[][];
+
+    // Validate data
+    if (!Array.isArray(data) || data.length === 0) {
+      throw new Error("Invalid or empty price data from CoinGecko");
+    }
+
+    // Data format: [[timestamp, open, high, low, close], ...]
+    const priceHistory = data.map((item: number[]) => ({
+      timestamp: item[0],
+      open: item[1],
+      high: item[2],
+      low: item[3],
+      close: item[4],
+    }));
+
+    // Calculate statistics
+    const prices = priceHistory.map(
+      (p: { timestamp: number; open: number; high: number; low: number; close: number }) => p.close,
+    );
+
+    if (prices.length === 0) {
+      throw new Error("No price data available");
+    }
+
+    const currentPrice = prices[prices.length - 1];
+    const dayStartPrice = prices[0];
+    const highPrice = Math.max(...prices);
+    const lowPrice = Math.min(...prices);
+    const priceChange = currentPrice - dayStartPrice;
+    const priceChangePercent = ((priceChange / dayStartPrice) * 100).toFixed(2);
+
+    return {
+      symbol: "SOL",
+      name: "Solana",
+      currentPrice: currentPrice,
+      priceChange: priceChange,
+      priceChangePercent: priceChangePercent,
+      high24h: highPrice,
+      low24h: lowPrice,
+      priceHistory: priceHistory,
+      dataPoints: priceHistory.length,
+      timestamp: new Date().toISOString(),
+    };
+  } catch (error) {
+    console.error("Error fetching SOL price history:", error);
+    throw error;
+  }
+}
+
 // Global Express error handler (must be last middleware)
 app.use((err: Error, req: Request, res: Response, next: (err: Error) => void) => {
   console.error("❌ Express error handler caught:", err);
@@ -559,14 +633,21 @@ app.use((err: Error, req: Request, res: Response, next: (err: Error) => void) =>
   });
 });
 
-// Start server
-app.listen(PORT, () => {
+// Start server with increased header size limit for Solana transactions
+const server = http.createServer(
+  {
+    maxHeaderSize: 32768, // 32KB - enough for large Solana transaction headers
+  },
+  app,
+);
+
+server.listen(PORT, () => {
   console.log("\n🚀 AnySpend Express Server with CoinGecko Premium Data");
   console.log("========================================================");
   console.log(`   Server running on: http://localhost:${PORT}`);
   console.log(`   Facilitator URL: ${FACILITATOR_URL}`);
   console.log(`   Network: ${NETWORK}`);
-  console.log(`   Payment USDC Amount: 0.001 USDC (1 * 10^6) to api/usdc/premium`);
+  console.log(`   Payment USDC Amount: 0.001 USDC (1000) to api/usdc/premium`);
   console.log(`   Payment B3 Amount: 100 B3 tokens (100 * 10^18) to api/b3/premium`);
   console.log(`   Payment for BTC: 0.01 USDC (10000 * 10^-6) to api/btc`);
   console.log(`   Pay To Address: ${PAYTO_ADDRESS}`);
@@ -579,7 +660,7 @@ app.listen(PORT, () => {
   console.log("   POST /api/btc               - BTC premium data (0.01 USDC)");
   console.log("   POST /api/solana/premium    - Solana premium (0.01 USDC on Solana)");
   console.log("\n💎 Premium Data Includes:");
-  console.log("   • 24-hour ETH/BTC price history (OHLC data)");
+  console.log("   • 24-hour ETH/BTC/SOL price history (OHLC data)");
   console.log("   • Current price & price change");
   console.log("   • 24h high/low prices");
   console.log("   • Historical data points");
@@ -589,6 +670,7 @@ app.listen(PORT, () => {
   console.log(`   • USDC Mint: ${SOLANA_USDC_MINT}`);
   console.log(`   • Payment Amount: ${SOLANA_PAYMENT_AMOUNT} (0.01 USDC)`);
   console.log(`   • Pay To Address: ${SOLANA_PAYTO_ADDRESS}`);
+  console.log(`   • Max Header Size: 32KB (for Solana transactions)`);
   console.log("\n💡 To test:");
   console.log("   Use the React client at http://localhost:3000\n");
 });
