+++ typescript/packages/x402-solana-wallet-adapter/package.json
@@ -36,6 +36,7 @@
   },
   "devDependencies": {
     "@types/node": "^22.13.4",
+    "@types/react": "^18.0.0",
     "@eslint/js": "^9.24.0",
     "eslint": "^9.24.0",
     "eslint-plugin-jsdoc": "^50.6.9",
@@ -44,18 +45,26 @@
     "@typescript-eslint/parser": "^8.29.1",
     "eslint-plugin-import": "^2.31.0",
     "prettier": "3.5.2",
+    "react": "^18.0.0",
     "tsup": "^8.4.0",
     "tsx": "^4.19.2",
     "typescript": "^5.7.3",
+    "@solana/kit": "^2.3.0",
     "@solana/wallet-adapter-base": "^0.9.23",
+    "@solana/wallet-adapter-react": "^0.15.0",
+    "@solana/wallet-adapter-react-ui": "^0.9.0",
     "@solana/web3.js": "^1.98.4"
   },
   "dependencies": {
     "@b3dotfun/anyspend-x402": "workspace:^"
   },
   "peerDependencies": {
+    "@solana/kit": "^2.3.0",
     "@solana/wallet-adapter-base": "^0.9.0",
-    "@solana/web3.js": "^1.90.0"
+    "@solana/wallet-adapter-react": "^0.15.0",
+    "@solana/wallet-adapter-react-ui": "^0.9.0",
+    "@solana/web3.js": "^1.90.0",
+    "react": "^18.0.0"
   },
   "exports": {
     ".": {
